+++ app/pianos/PianoWithChordsCheatsheet/PianoWithChordsCheatsheet.css
