--- app/routes/ChordsHelperPage.tsx
@@ -1,10 +1,0 @@
-import React, { useState } from "react";
-import PianoWithChordsHelper from "../pianos/PianoWithChordsHelper/PianoWithChordsHelper";
-
-export default function ChordsHelperPage() {
-  return (
-    <>
-      <PianoWithChordsHelper />
-    </>
-  );
-}
