+++ app/routes/PianoChordsCheatsheetPage.tsx
@@ -1,0 +1,10 @@
+import React, { useState } from "react";
+import PianoWithChordsCheatsheet from "../pianos/PianoWithChordsCheatsheet/PianoWithChordsCheatsheet";
+
+export default function ChordsHelperPage() {
+  return (
+    <>
+      <PianoWithChordsCheatsheet />
+    </>
+  );
+}
